refactor: 이름과 폴더구조 변경
- port는 ports폴더로 이동
- 이름에 catchMind가 필요없는 파일 이름 변경
