+++ packages/backend/src/domains/catchMind/controllers/catchMindAPI.controller.ts
@@ -1,11 +1,11 @@
-import { CatchMindInputPort } from "../useCases/CatchMindInput.port";
-import { CatchMindService } from "../useCases/catchMind.service";
+import { CatchMindControllerPort } from "../useCases/ports/CatchMind.controller.port";
+import { CatchMindService } from "../useCases/catchMind.useCase";
 
 import { DomainConnecter } from "../../../utils/domainConnecter";
 import { StartData } from "../../../types/catchMind.type";
 
 const connecter = DomainConnecter.getInstance();
-const gameService: CatchMindInputPort = new CatchMindService();
+const gameService: CatchMindControllerPort = new CatchMindService();
 
 connecter.register(
   "catchMind/game-start",
