+++ packages/backend/src/types/catchMind.type.ts
@@ -1,5 +1,23 @@
 import { Player } from "../domains/catchMind/entity/player";
 
+export type CatchMindInfo = {
+  roundTime: number;
+  currentRound: number;
+  turnPlayer: string;
+};
+
+export type StartGameData = {
+  totalRound: number;
+  roundInfo: CatchMindInfo;
+};
+
+export type RoundEndData = {
+  roundWinner: string | null;
+  suggestedWord: string;
+  playerScoreMap: { [K: string]: number };
+  isLastRound: boolean;
+};
+
 export type StartData = {
   goalScore: number;
   players: string[];
